Added Menu
Added a basic Drop down/side scroll menu. did a sneaky little trick to make the menu come out on click. I put a check box on top of the hamburger menu icon and set the opacity to 0 so it cant be seen. When someone clicks on the button it opens menu.